+++ src/context.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 let defaultOptions = {
-  bindI18n: 'languageChanging languageChanged',
+  bindI18n: 'languageChanged',
   bindI18nStore: '',
   // nsMode: 'fallback' // loop through all namespaces given to hook, HOC, render prop for key lookup
   transEmptyNodeValue: '',